dry refactor cardlistitem
