+++ src/components/CardList.vue
@@ -7,6 +7,7 @@ import { useSpaceStore } from '@/stores/useSpaceStore'
 import { useThemeStore } from '@/stores/useThemeStore'
 import { useGlobalStore } from '@/stores/useGlobalStore'
 
+import CardListItem from '@/components/CardListItem.vue'
 import UserLabelInline from '@/components/UserLabelInline.vue'
 import ItemCheckboxButton from '@/components/ItemCheckboxButton.vue'
 import NameSegment from '@/components/NameSegment.vue'
@@ -65,7 +66,7 @@ const props = defineProps({
   dateIsCreatedAt: Boolean,
   resultsSectionHeight: Number,
   currentCard: Object,
-  showSpaces: Boolean
+  disableListOptimizations: Boolean
 })
 
 const state = reactive({
@@ -91,73 +92,11 @@ const normalizedCards = computed(() => {
     return card
   })
 })
-const urlPreviewImage = (card) => {
-  if (!card.urlPreviewIsVisible) { return }
-  return card.urlPreviewImage
-}
 const selectCard = (card) => {
   emit('selectCard', card)
 }
 const removeCard = (card) => {
   emit('removeCard', card)
-}
-const cardIsActive = (card) => {
-  const isCardDetailsVisible = globalStore.cardDetailsIsVisibleForCardId === card.id
-  return isCardDetailsVisible || card.isLoading || isCurrentCard(card)
-}
-const cardIsFocused = (card) => {
-  return globalStore.previousResultItem.id === card.id
-}
-const cardDate = (card) => {
-  return props.dateIsCreatedAt || card.nameUpdatedAt || card.updatedAt
-}
-const relativeDate = (card) => {
-  const date = cardDate(card)
-  return utils.shortRelativeTime(date)
-}
-const userIsNotCurrentUser = (userId) => {
-  return userStore.id !== userId
-}
-const isStrikeThrough = (card) => {
-  return card.name.startsWith('[x]')
-}
-const isThemeDark = computed(() => themeStore.getIsThemeDark)
-const colorIsDark = (card) => {
-  if (!card.backgroundColor) {
-    return isThemeDark.value
-  }
-  return utils.colorIsDark(card.backgroundColor)
-}
-const dateIsToday = (card) => {
-  const date = cardDate(card)
-  if (!date) { return }
-  return dayjs(date).isToday()
-}
-const isCurrentCard = (card) => {
-  return props.currentCard?.id === card.id
-}
-const nameIsColor = (card) => {
-  const name = card.name.trim()
-  return colord(name).isValid()
-}
-const cardInfoStyles = (card) => {
-  const styles = {}
-  if (card.backgroundColor) {
-    styles.backgroundColor = card.backgroundColor
-  }
-  if (nameIsColor(card)) {
-    styles.backgroundColor = card.name
-  }
-  return styles
-}
-
-// [·] checkbox cards
-
-const cardIsTodo = (card) => {
-  return Boolean(utils.checkboxFromString(card.name))
-}
-const canEditCard = (card) => {
-  return userStore.getUserCanEditCard(card)
 }
 
 // scroll
@@ -211,30 +150,16 @@ span
   ul.results-list.card-list(ref="resultsListElement")
     .prev-scroll-area-height(:style="{height: state.prevScrollAreaHeight + 'px'}")
     template(v-for="card in itemsRendered" :key="card.id")
-      li(@click.stop="selectCard(card)" :data-card-id="card.id" :class="{active: cardIsActive(card), hover: cardIsFocused(card)}")
-        //- date
-        span.badge.status.inline-badge(v-if="!props.shouldHideDate" :class="{'date-is-today': dateIsToday(card)}")
-          img.icon.time(src="@/assets/time.svg")
-          span {{ relativeDate(card) }}
-        //- space
-        span.badge.status.inline-badge(v-if="card.space")
-          span {{ card.space.name }}
-        //- user
-        UserLabelInline(v-if="card.user.id && userIsNotCurrentUser(card.user.id)" :user="card.user")
-        //- [·]
-        ItemCheckboxButton(:visible="cardIsTodo(card)" :card="card" :canEditItem="canEditCard(card)" :parentIsList="true")
-        //- card info
-        span.card-info(:class="{ badge: card.backgroundColor || nameIsColor(card), 'is-dark': colorIsDark(card) }" :style="cardInfoStyles(card)")
-          //- name
-          template(v-for="segment in card.nameSegments")
-            img.card-image(v-if="segment.isImage" :src="segment.url")
-            img.card-image(v-if="urlPreviewImage(card)" :src="urlPreviewImage(card)")
-            NameSegment(:segment="segment" :search="props.search" :isStrikeThrough="isStrikeThrough(card)" :backgroundColorIsDark="colorIsDark(card)")
-          //- remove
-          button.small-button.remove-button.danger(v-if="props.cardsShowRemoveButton" @click.left.stop="removeCard(card)")
-            img.icon(src="@/assets/remove.svg")
-          //- loading
-          Loader(:visible="card.isLoading" :isSmall="true")
+      CardListItem(
+        :card="card"
+        :search="props.search"
+        :cardsShowRemoveButton="props.cardsShowRemoveButton"
+        :shouldHideDate="props.shouldHideDate"
+        :dateIsCreatedAt="props.dateIsCreatedAt"
+        :resultsSectionHeight="props.resultsSectionHeight"
+        :currentCard="props.currentCard"
+        @selectCard="selectCard"
+      )
 </template>
 
 <style lang="stylus">
